Move tools to main folder
