+++ README.md
@@ -9,7 +9,7 @@ Create Pixel Art games in Golang with fun and ease.
 ## What you can do with Pixiq?
 
 + draw images on a screen in real time using your favourite [Go programming language](https://golang.org/)
-+ manipulate every single pixel directly or with the use of tools (_tools are not ready yet_)
++ manipulate every single pixel directly or with the use of tools
 + handle user input (_only keyboard supported at the moment_)
 
 ## What is Pixel Art?
