+++ clear/clear.go
