+++ clear/clear_bench_test.go
@@ -3,9 +3,9 @@ package clear_test
 import (
 	"testing"
 
+	"github.com/jacekolszak/pixiq/clear"
 	"github.com/jacekolszak/pixiq/image"
 	"github.com/jacekolszak/pixiq/image/fake"
-	"github.com/jacekolszak/pixiq/tools/clear"
 )
 
 func BenchmarkTool_Clear(b *testing.B) {
