+++ clear/clear_test.go
@@ -1,45 +1,24 @@
-package glclear_test
+package clear_test
 
 import (
 	"fmt"
-	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
-	"github.com/jacekolszak/pixiq/glfw"
+	"github.com/jacekolszak/pixiq/clear"
 	"github.com/jacekolszak/pixiq/image"
-	"github.com/jacekolszak/pixiq/tools/glclear"
+	"github.com/jacekolszak/pixiq/image/fake"
 )
 
-var mainThreadLoop *glfw.MainThreadLoop
-
-func TestMain(m *testing.M) {
-	var exit int
-	glfw.StartMainThreadLoop(func(main *glfw.MainThreadLoop) {
-		mainThreadLoop = main
-		exit = m.Run()
-	})
-	os.Exit(exit)
-}
-func TestNew_Integration(t *testing.T) {
+func TestNew(t *testing.T) {
 	t.Run("should create tool", func(t *testing.T) {
-		openGL, err := glfw.NewOpenGL(mainThreadLoop)
-		require.NoError(t, err)
-		defer openGL.Destroy()
-		command := openGL.Context().NewClearCommand()
-		// when
-		tool := glclear.New(command)
+		tool := clear.New()
 		assert.NotNil(t, tool)
 	})
 }
 
-func TestTool_Clear(t *testing.T) {
-	openGL, err := glfw.NewOpenGL(mainThreadLoop)
-	require.NoError(t, err)
-	defer openGL.Destroy()
-	command := openGL.Context().NewClearCommand()
+func TestClear(t *testing.T) {
 	t.Run("should clear selection", func(t *testing.T) {
 		colors := []image.Color{
 			image.RGBA(10, 20, 30, 40),
@@ -75,9 +54,9 @@ func TestTool_Clear(t *testing.T) {
 			for name, test := range tests {
 				testName := fmt.Sprintf("%s %v", name, color)
 				t.Run(testName, func(t *testing.T) {
-					img := openGL.NewImage(2, 2)
+					img := image.New(fake.NewAcceleratedImage(2, 2))
 					selection := img.Selection(0, 0).WithSize(test.width, test.height)
-					tool := glclear.New(command)
+					tool := clear.New()
 					tool.SetColor(color)
 					// when
 					tool.Clear(selection)
@@ -91,8 +70,8 @@ func TestTool_Clear(t *testing.T) {
 
 func assertColors(t *testing.T, img *image.Image, expectedColorLines [2][2]image.Color) {
 	selection := img.WholeImageSelection()
-	assert.Equal(t, expectedColorLines[0][0], selection.Color(0, 0))
-	assert.Equal(t, expectedColorLines[0][1], selection.Color(1, 0))
-	assert.Equal(t, expectedColorLines[1][0], selection.Color(0, 1))
-	assert.Equal(t, expectedColorLines[1][1], selection.Color(1, 1))
+	assert.Equal(t, expectedColorLines[0][0], selection.Color(0, 0), "position(0,0)")
+	assert.Equal(t, expectedColorLines[0][1], selection.Color(1, 0), "position(1,0)")
+	assert.Equal(t, expectedColorLines[1][0], selection.Color(0, 1), "position(0,1)")
+	assert.Equal(t, expectedColorLines[1][1], selection.Color(1, 1), "position(1,1)")
 }
