+++ glblend/integration/glfw/integration_bench_test.go
@@ -3,9 +3,9 @@ package glfw_test
 import (
 	"testing"
 
+	"github.com/jacekolszak/pixiq/glblend"
 	"github.com/jacekolszak/pixiq/glfw"
 	"github.com/jacekolszak/pixiq/image"
-	"github.com/jacekolszak/pixiq/tools/glblend"
 )
 
 var resolutions = map[string]struct {
