+++ glclear/glclear.go
